+++ R/fonts.r
@@ -3,6 +3,5 @@
 #' 360info fonts without having them installed.
 fontfaces_360fonts <- function() {
   list(paste0(
-    "https://fonts.googleapis.com/css2?family=Public+Sans:",
-    "ital,wght@0,400;0,700;0,900;1,400&display=swap"))
+    "https://fonts.googleapis.com/css2?family=Roboto+Condensed:ital,wght@0,300;0,700;0,800;1,300;1,700&display=swap"))
 }